Add more docstrings to functions. Add type hints to class attributes and correct the docstrings for class attributes so that sphinx.autodoc will capture it.
In case of limiter configurations, the function that aligns the
equilibrium LCFS with the target geometry is simplifed, basically
the shortest distance between the part of the geometry that is the
closest to the LCFS is considered as the adjustment.

Improve the JSON loading function, by first removing C style comments
from a JSON content and remove trailing commas.